working game
picture + wordle game matched.
page will refresh on win or lose
