+++ src/app/Components/Wordle.css
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   gap: 20px;
+  padding-top: 20%;
 }
 .Wordle-form {
   display: grid;
