+++ src/app/components/picture.tsx
@@ -1,10 +1,17 @@
+"use client";
+
 import Image from "next/image";
 import Chat from "./chat";
 
+import { useContext } from "react";
+
+import { PokemonContext } from "./PokemonContext";
+
 import "./picture.css";
+
 export default function Picture() {
-  const randomPokemonNumber = Math.floor(Math.random() * 151);
-  const poke = randomPokemonNumber.toString().padStart(3, "0");
+  const contextData = useContext(PokemonContext);
+  const poke = contextData.pokemonID.toString().padStart(3, "0");
   const randomBackgroundNumber = Math.floor(Math.random() * 23 + 1);
   return (
     <div className="outerBox">
